move multi picker to folder
