+++ lib/screens/stats/stats.dart
@@ -1,7 +1,7 @@
 import 'package:budgetiser/drawer.dart';
 import 'package:budgetiser/screens/stats/line_chart.dart';
 import 'package:budgetiser/shared/dataClasses/account.dart';
-import 'package:budgetiser/shared/picker/account_picker.dart';
+import 'package:budgetiser/shared/picker/multi_picker/account_picker.dart';
 import 'package:budgetiser/shared/picker/month_picker.dart';
 import 'package:flutter/material.dart';
